Change unzip_archive output to file
Instead of a directory, the output of unzip_archive is now a
supplementary file. That way, unzip_archive isn't accidentally called
when the unzipped files inside a directory are needed. However, this
doesn't fix the underlying issue that unzip_archive doesn't get called
for dynamic input files when pull_date.txt is changed.